cleanup(angular): add @angular-eslint/builder:lint builder migrator
