+++ packages/angular/src/generators/ng-add/migrators/builders/index.ts
@@ -1,4 +1,5 @@
 export * from './angular-devkit-karma.migrator';
 export * from './angular-devkit-ng-packagr.migrator';
+export * from './angular-eslint-lint.migrator';
 export * from './builder-migrator-class.type';
 export * from './builder.migrator';
